Changes to Vox clothing and storage
The clothing for vox had not been updated when the changes to accessorys and how they attached were made.
This has updated vox under clothing, and the vox storage harness to make use of the standard layout

Specificity it removed the unique "vox" accessory point, which was not defined, therefor didn't work (so you couldn't equip the vox webbing to vox clothing, even).
And made them use the same valid/restricted points as their parents.

This does however mean that vox under suits can equip things as per a standard jumpsuit could be expected to, and has no special rules associated anymore for been..well..vox
Same can be said for the webbing, it can be equipped to anywhere as per a normal webbing rules, and has no specific restrictions anymore simply for been vox-made.