Blueshield update/oversight
So..blueshield in loadout wise is a mix of command and security gear.
So..i've (finally) added Blueshield Guard to anything that was restricted to either "security officer", or to the 'entire list of command staff'.
this allows BSG's to take holsters, Sec huds, and the PMC coats and all that, along with coats that are usually in sec/command only for loadout choices..

What this doesn't mean is you can take the hat that only the CD or HoS has. those are still locked. I specifically  looked for the two above criteria for access restrictions.

This is something that was long due and forgotten about in the midst of everything else.