Adds ability to lick ones own wounds with the new wound-lick ability
I won't lie, this is a possibly slightly hacky way around it, by changing the range check from the living_mob(1) in the var itself, to an if check later in.

What this means in the front side, is that all carbon/human mobs on screen will show up in the targetable list if its used instead of just right clicking someone, and the 'check' for range will happen just before licks happen. telling them to get closer to the intended person instead.

if theres a cleaner way of doing this, of which I'm fairly sure there is, we can change to that when we find it.