Suit Cycler Armor Fix - Down/Upstream Earlyport
This is a port of https://github.com/PolarisSS13/Polaris/pull/6642 which is a port of https://github.com/VOREStation/VOREStation/pull/6483 which is a port of https://github.com/CHOMPstation/CHOMPstation/pull/695.

PORTS GALORE.

These apply the correct stats when placing a voidsuit in a cycler and switching from, say, engineering to atmos.