Changing blob allowances so blob types can correctly use blob buildings.
Player blobs had hard-coded restrictions for some buildings, preventing all blobs types from using them. This edit changes that, so the blob type restrictions are what prevent a blob from making a building it should not be able to, not a hard coded restriction for all blobs. Nodes, Factories, and Resource buildings can now be used and made by a player blob if their blob type allows it. There are no known balance issues with these buildings, nor any known bugs. The blob resource building was already pre-balanced for a MRP/HRP server, so no action needs to be taken for balancing.