Makes the teleporter stop killing you all the time.
That prob(5) is the defining reason why nobody ever wants to use the teleporter.
It's supposed to be called when the teleporter is inaccurate, such as back when teleporter calibration was a thing. But the teleporter doesn't do calibration anymore, so that accuracy var is always 0.

Stop throwing command staff into space plz.