Activates AI multi-cam on Cryogaia.
As in the title.
Adds a multicam location for AI's so the function will work. tested as AI and checked it all works nicely. so far so good.
This area is set up to the right of the ship map in centcomm maps.
hopefully this wont interfere with any of it, if so we can just..nudge it further east or something if need be.