Adds Spice Resistence/Intolerance Traits, Makes Frost Oil COLD
- Spice Resistance traits in the same flavor as Booze traits. 
- 3x, 2x, 1.5x, 0.75x, 0.5x, 0.25x resistance to the effects of capsaicin and frost oil.
- Frost Oil and Capsaicin will now **lower/increase your body temperature by a random amount** during the ingest step if you're over the danger threshold.
- var/spice_mod is added to the species datums, much like alcohol mod.
- Frost Oil will now remove 5 capsaicin on affect_ingest much like capsaicin would remove 5 frostoil.

Not much else to say.
I got tired of keeling over in agony after eating spicy food, and frankly, this is an RP server, so mechanical hardlimits on characters that canonically power down spicy food without giving a shit is frustrating to say the least.

Earlyport of: https://github.com/VOREStation/VOREStation/pull/8726