Clear Skies Ahead!
Clears up a hell of a lot of clutter with all of the pre-defined areas of shuttles to properly use the landmarks system. Now there's a ton of free room to build and/or allow larger shuttles to dock with the station, as well as allowing overmap shuttles to dock at the docks.