Changing Sickshot desc and info_desc
The Sickshot states in its description that it can stun, and because of that implies that it is a self defense weapon. It is not a self defense weapon, as it only does 5 burn damage and causes some very minor confusion, not stuns. Its true use is to forcibly expel the contents of stomachs, so the description has been changed to make that a bit more obvious.

Also for some silly reason the sickshot doesn't live in a normal weapon file, so I have to edit the sickshot_vr to make this work. Not much I can do about that.