ranged_cooldown_time variable for mobs
This is the addition of a new mob variable that will allow mobs to move and shoot at the same time. This section here is special snowflake code for metroids only, or for whatever else in the future that you want to have move and shoot at the same time. Basically, this is a non-stupid version of "ranged_attack_delay." "ranged_attack_delay" is stupid because it sleeps the entire mob. This new ranged_cooldown_time is smarter in the sense that it is an internalized timer. Try not to confuse the names.