Remove shit opinion
Removed a stanza that prevented some flexibility with the digestion for mobs, and also the returned comment was highly unprofessional.
You can now toggle the digestion mode on simple animals. You will be warned that using this to break OOC prefs is against the rules before doing so. Right click the mob and toggle their digestion.

It's VORE it's not meant to be balanced anyway.