Remove transit from SC extra_z_levels
This is somehow causing people to drift into the transit level. I guess initially this was done so that radio comms can make it to the transit level. I don't know how to test that on a private server, so I am putting it to main and hopping for the best, heheh.