Give carp event a 50% chance to even occur.
If the carp event procs, there is a 50% chance for it to not occur, giving a free moment of peace for that event timeslot.

There are a lot of security events already including the minor carp invasion, major carp invasion, drones, blob, spiders, and metroids. That is a lot to handle. So this is part of a slight effort to tone that down.