New Wheel of Fortune
Made a new grander and more attention grabbing wheel of fortune sprite.
Remade code so now it does public spins, and it handles sale of lottery tickets itself. Staff with right access can enable/disable public spin and sales, reset lottery and of course do a roll of the wheel for the lottery

Good luck everyone!