Fix loading admin ranks if NOT using legacy system
Error found by Cadyn. We're using a database to keep track of admins, so we are using the database system, but the database system does not load from the admin_ranks file, so no admin ranks are being loaded at all.

Decided to move it out of the legacy system if statement and just force it to load before deciding if its legacy or database.