Adds Temperature Adaptations to Xenochimera along with Autohiss.
Ran these by Scree (Creator of Xenochimera) and he said these are fairly reasonable adaptations.

They're neutral traits available to custom species, and given that Xenochimera are also able to pretty much mix and match a lot of stuff based on what they've eaten/evolved, this will allow greater roleplay/adaptations to occur - for instance, a xenochimera that is well-suited to the warmth of the desert is going to struggle on Sif, and vice versa.