Blueshield Guard Revival
This is gonna generate a lot of mixed feelings from people, from what I'm already seeing.

I decided to go through and move Blueshield to _ch and such files to prevent data loss if we ever fully cleanse the _yw files remaining in the codebase, partly also because I looked through PRs and realized - Blueshield's been +gone+ for well over two years.

I figure by now, maybe things have changed enough that they might be trusted if added to the whitelist that all Command has to pass. Figured I'd at least throw up the PR and see what staff had to say.

Some questions people are probably going to ask:
1: Do I (Rykka/Sonya/etc) think Blueshield belong here?
- Eh? They've got a role that overlaps with Security - protecting the Heads, and particularly the CD. There's potential for sec overlap and a lot of head-butting, but it'd be neat for the event rounds.
2: What if Command wants to get eaten?
- Then Blueshield shouldn't interfere if it's scene non-con. Same as Sec ought to not interfere unless it becomes public. In this case, BS ought to just "accidentally" wander off to another room during Extended. During event rounds? Yeah, defend your CD, imo.
3: How should a role like this play during Extended?
- The same as if HoS and CommandSec were rolled together, almost, except tightly-glued to Command. Chill with your boss, keep them from doing stupid shit like waltzing off into the wilderness, etc etc.

Please try to keep feedback focused solely on the role and how it'll interact with the server. If you had a bad experience with a Blueshield officer - I'd like to hear it and hear concerns on how to address it, but saying all Blueshield officers are awful may not be usable without further specifics as to WHY.