Enables all metabolism types for all species
Fast, Slow, Apex, and Singularity metabolisms are now available for all species/characters irregardless of if custom species or not.

This shouldn't affect gameplay balance in any way, EXCEPT potentially on 'chimera - therefore it's advised that, as with the Neural Hypersensitivity trait for Xenochimera, you do not take it unless you understand what you're doing and WILL play around it.

Singularity Metabolism was renamed Metabolism, Singularity in the trait display list to line up better with the other metabolism traits.