Possibly fixes some spiderling jank
Ran out of brain juice trying to figure out the bugginess further but the wacky placement of this stuff got me sussy enough to try something. will see if anything happens and maybe toss upstream if so.

If my hunch is correct this might fix excess unregistered spiderling spawning and the bug with infinitely squishable undeleted squishees.