Update phazon.dm
Adjusted the number of slots that both phazons have, making the Janus better and the normal phazon worse.  Bumped up the health of the normal phazon by 50 to 250 and removed its ability to cloak, which will be unique to the Janus and Scuttlebug.  This makes the phazon a solid tier 4 mech, and makes it not overpowered for something science can build.  
Will change the armor later, as I intend to change a lot of those.