Breathing Sounds while in Space
Plays a breathing sound akin to Bay's breathing sound while you're in space.

Does not come with a pref currently, as it's quiet - and intended to be an audible indicator of if you're breathing or not.
(IE if you don't hear it, be concerned.)

Will be ported upstream in a moment, hence the lack of CHOMPEdits.