Reduces Cryo timer to 60 from 599, in-line with other round-exit methods
This will fix leaving via cryostorage and being able to join via other methods to delete yourself via manifest, intentionally or accidentally.

Also re-converts SC 1 and 3 to TGM format.
Please use SDMM in the future (The launcher is bundled with the codebase), or run dmm2tgm.bat in /tools/mapmerge2 if using DreamMaker.