Reduce loneliness and major loneliness points
Loneliness -3 --> -1
Major loneliness -5 --> -3

Why: Loneliness was a source of a lot of free points for some builds. Adventuring into the wilderness with a partner allowed for a major points advantage where having a buddy already supplemented the crew's advantage. Also this is being reduced in light of removal of the negative traits cap.