Surgery QOL: Makes object removal take time rather than a pass/fail.
Rather than prob 60/40/50 in the case of legal implant, illegal implant, and shrapnel/other in that order, surgery now has a time-to-remove.

Legal implants take 10 seconds, illegal take 20 (to represent you needing to dig around/hunt for it), and shrapnel/other stuff takes 10 seconds.

This will make surgery for shrapnel removal much less of a PITA, and should be more fun for players now as a result. Rather than YOU JUST MISSED IT THIS TIME x72.