Chapel POI edit + Gravekeeper Drone Bugfix
Added some goodies to the chapel to make it a rest stop for explorers, and a viable base of operations for the gravekeeper drone. Increased it's generator recharge rates and gave it a SMES. If you manage to blow through all that power now...

Fixed the gravekeeper drone's emag item. Apparently, they only get one. Oops. Also gave it the ability to actually clean messes.