Engineering QoL
Fixed a bug with the shield generators in Misc Research.

Major edits to the engine room to allow (and encourage) extra TEG's to be set up post engine startup.

Wired up a shieldgen SMES. It is off by default, and the shields remain in storage.

Mirrored #4275 to avoid conflicts. Also needed to be done.

Adjusted waste gas pipes a bit.

Fixed a loading decal direction in kitchen.

Swapped the Storage tank location with the SMES crate for ease of access.