Add missing sprites to teshari for a few misc items.
A few items were missing sprites for teshari, and for my first PR I decided to give a few of the miscellaneous items that were already in the game proper sprites to fit teshari. 
https://i.imgur.com/ihCMzVM.png

Though with the emotional mask, there was only sprites fitting for humans and shadekin, no sprites for the other species IE: Vulpakin, Unathi, etcetc

This is my first code go easy on me, and I'm still not sure if this is how the modular system works.
Code compiled and all that.