Fixes the Ursula's wiring and gives the Explo Sling bluespace flares.
Currently the wiring for the Ursula is completely nonfunctional, with the SMES's input and output being connected, this easy fix should make it a functional ship now.

The SMES now outputs default now but the APC is still off by default so engineers don't get mad over power alarms. Bring a crowbar.

I kept the ship disconnected so the powergrid of the sling's limited 200kw doesn't get stressed, bring gloves.

Bluespace flares are mapped in because despite being a science printable, people are unaware it exists, this flare will enable you to land the overmap shuttles on Sif and other places if there's enough space. They are found in the explo prep room.