Makes botany less depressing, adds a art gallery, and moves the library
This might be a more controversial change as it completely moves entire departments around.

CHANGES: 
Botany is much larger now, gardeners don't have to get crammed in 1 tile walk ways in their dark barely lit botany rooms.

The trays are intentionally placed so people can see every tray from the window and the kitchen, gives the botany a bigger public eye instead of being closed off unlike the rest of the service departments AND allows someone to chef and botany without any plants off screen. 10 trays in sight 

There is now an empty animal pen so when someone orders chickens or cattle, you don't have to put them on the floor or kitchen.

Added a art gallery in place of the library too complete with art supplies and canvases, including painting frames that might be potentially PERSISTANT between shifts

More painting frames around the station might be a separate PR

These changes got in the way of the coffee shop and library so I moved both of those things in the third deck REPLACING the public garden that no one used.

The xenobotany seeds that were there are left unchanged, the center ring already had public dirt piles for plant growing for bored players with equipment.

The library is now in place of the public garden, still complete with a cafe!

Minor mapping change on the large public bathroom is accessible from the locker room area rather then the library...

Disposals has their own cargo mailing area now known as "Art Gallery" along with having the library one re-sorted. Thanks TheGreatKitsune!

Why its good for the game?
Service is less depressing and bare-bones then how it currently is, the art gallery gives visitors and players a thing to do when there is down time. Botany is in more public eye to open up for more RP.

Known issues:
Because of how air alarms and open spaces caused issues with the public garden, the APC was in the cafe and it is a shared APC. It was always like this but that means light switches in the cafe and library are shared too.