Defib Buff
Removes the mental interface error on defibs.

Nothing is more annoying for players, medical and patient alike, for an OOC issue to bleed into an in-character one.

A brief bout of afk frustration or loss of internet leading medical to try and revive you using many chems, then giving up an leaving you for dead... A Search and Rescue going out to rescue someone in such a state... Why punish players of medical for doing their job with a no-sale at the absolute last step?

Bonus points, monkey's are now revivable. Yay!