Updates sprite location for new Suspension Field Generator FIXES #4971
Updates the location for the sprites, including the one for when the device is on and wrenched down.

Also fixes an issue where the user could open the maintenance hatch, lock the device, then close the TGUI window. This would prevent them from being able to open the TGUI window to unlock it and they would get an access denied message when trying to close the hatch.