Character Directory Update - Adds Four Additional Tags
This will be ported upstream to VOREStation momentarily, hence the lack of // CHOMPStation comments.
If upstream makes any changes or rejects the PR, I will add comments on as needed.

Adds four new preferences + tags, shown here!
![](https://cdn.discordapp.com/attachments/404660884531707906/1041243283373948978/azcYOMj.png)

Gender, Sexuality, BDSM Role, and Furry/Human Preference.

The latter two may be unnecessary, particularly the BDSM role, but I figured it was an easy addition, and just as easy to remove on request!

The intent for these additions is to make perusing for scenes ever-so-slightly easier, and lay out important keynotes such as gender more clearly.