Doom Peninsula Overhaul
Expanded the merc base upon the peninsula, giving it a purpose of being a research base. Added some more goodies, and more foes. Turrets might be a tad overboard.

Oh and added 8? gyropistol clips, and a cell battery to the manor. Cell battery is in the storage room, and 3 clips with the gyropistol, while 5 are in a closest in another room. Note gyropistol cannot be made. Those 8 clips are all you're getting.