Fixes Discomfort Messages not showing up at all
If Custom Cold + Custom Heat are not set on a mob, the list is empty and for w/e reason doesn't initialize properly, resulting in a null value. This is only resolved when you reset the list first, but even then, it still remains null in-game. This is a fix in case the custom_ var is ever null, it will fallback to the default, so as to not break thrown alerts.

Fix for the ACTUAL cause (variable being null and list not initializing, resulting in users hitting "edit" on the custom strings in menu and it not working until it's reset first at least one TBD).