Fix Supermatter Warning the Entire Station on explode
SM, as it exists currently, will play the delamination warning over radio comms/globally when it is close to exploding, even if it isn't located on the station.
It would also benefit from a causality field, even outside it.

This fixes that.

Supermatter will no longer warn the entire station if it is presently being damaged unless it exists on the same levels as the station. 

It will also skip the causality field otherwise, though this is subject to Headmin review.