Spaceman DMM Proc Debugging
Allows you to debug via breakpoints when a runtime occurs, and chase the proc call stack in VSCode.

Run by hitting F5 and allowing the game to run from there, then when it hits a runtime, the game will pause, allowing you to view what caused/led up to the runtime.