Enables Testing Character Voices in Setup
As title;
This function was WIP, @blackmajor didn't have time to finish it.
I went ahead and finished this up, and realized it was worth PR'ing separately for speedmerge, because this is MASSIVELY convenient.

TL;DR:
You can hit "Test Character Voice" in setup, to play a preview of what your voice will sound like with the selected type and frequency.

See demonstration here: https://streamable.com/m6u9vc