Admin QOL Logs
Adds a log for Aghosted admins returning to their body, regardless of method.
Adds a log for Admins logging in to the game. This happens on client/Topic, so it works in 99.9% of cases and shouldn't spam (like when I tried it on /mob/Login)
Fixes a potential infinite loop in display_roundstart_logout_report() if there's no clients on the server at all.