Buffing Suppressor
Merc suppressor's threat seemed based off when velocity damage was a thing.
Damage from 10 to 30

Oh and also nerfed the base bullet from 60 o 15 because some bullets are using that damage value for some reason, and really, why the was it that high?
Looking in to trying to fix some broke guns.