Secondary Auto-Transfer
Adds in a secondary location and chance for auto transfers. It required reworking the logic regarding transfer queues with min and max transfer settings. When objects transfer they will individually attempt the primary location first, and failing that it will attempt the second. If both fail the object's transfer cooldown starts again.