Fixes proteans being immune to duffle chonk
Probably wouldn't have even noticed if it wasn't for some gamer deciding to flex this exploit with the double-duffle-slowdown taur saddlebags as non-taur of all things instead of just going with a regular duffle of same capacity.

Also why the f do the saddlebags even have 2x duffle slowdown in the first place? that's messed up.