CMO Equipment and Med Storage Additions
Adding three "cases" of medicine for the CMO to have at round start to give them something useful that is "tangible" to the station. The advanced hypo is... not really all that useful let's be real.
Also, everything that was redeleted from storage during a small medical change has been added back in, along with one addition for xenochimera. Two boxes of BrainSnax.