Bring Your Own Lunch To Work Day
Makes the lunchbox item free for all employees to bring to work. Honestly, if this is an issue, I'd still like to lower it to a cost of 1. It's a little neat flavour thing that can add to a character so people can bring in their "favourite meals" to work and such. And gives people a reason to stop by the breakrooms to put their lunch in their lockers and stuff. I think it's neat. But there's so many fluff and roleplay tools that it's hard to bring all you want. Implants, special breath masks, your own outfit from work, collars for those into it. And lunch costs two points. (Mainly this is an annoyance for folk who bring computers and such to work. Which I'm fine with them being expensive. It just adds up~)