Modular appearance for synx simplemobs
Moves the synx simplemob onto a modular, overlay-based system for sprites, fixes several bugs with their basic code, and adds sprites for their transform and distend stomach verbs. There's still some ground I'd like to cover with this and the customization options are a bit barebones for now, but the system itself is completely functional and will serve as a good basis to add future additions to. Also makes synx play a bit more accurately to the lore as well as part of some minor gameplay tweaks.