Confirmed outbreak of a less stupid biohazard system
I call it stupid but really this is just a result of really old Baycode that no one ever bothered to clean up or make consistent in any way so people have just been sort of blindly doing their own thing because nobody told them otherwise.

From now on.
Level 7 biohazard = stuff that's a danger but isn't immediately deadly.
Level 5 biohazard = more severe stuff that you ought to react to like right now.
The lower the number goes, the more urgent it is to handle. Level 7 is just the start of things that all crew have to react to.

Furthermore, I made "Biohazard Alert" actually tell you what the fucking alert is now. So no more stupid "Wtf is a level 5 again?" "It's virus" "No it's vines" "No it's wallrot." Turns out, the alerts were made up and the numbers didn't matter. I fixed that.

Vines no longer trigger a level 5 alert also.

In conclusion, this is not a good solution but it's better than what we had. Basically, the lower the number, the more urgent the issue is. A hypothetical level 1 biohazard would be just "the station is no man's land, arm the nuke, it must never be allowed to spread, nothing may survive, and quarantine the area afterwards just to be safe".

Really we should just redo literally all the AI voice lines, but that's out of the scope of this PR.