Poop water
this server is rotting my brain. this server yes this server and the shitcode. like metaphorically if this say let's say imagine my brain as a drawer of socks. a sockdrawer so to speak filled with socks. socks. socks and shitcode yes. now imagine a wildfire. but not in the wild. contained. to a drawer full of socks for example. a sockdrawer so to speak. i got it under control. the shitcode not the socks. those are still in the wild firedrawer. its all on fire just like my brain. on fire but under control and under socks. socks out in the wild. like where me and my mate paul used to to do drugs in our socks in. controlled drug use in a wild sockdrawer. it was wild. too wild. nothing out of the sockdrawer tho. thats where i stashed my drugs. in a wildire. an ordinary sockdrawer for drugs. gave me shitcode i fear the wildfire. but i got it under control. the socks in my shitcode. this server is rotting my brain. but i got it under control.