Desert Club SP-Pod
Sand enjoyers rejoice!

This addition to our list grants a purely social SP-Pod for Outsiders!

The pod is fitted with the following:

1 Bar
1 Kitchen
1 Botany Room
1 Storage Room
1 Bathroom
1 Dining Floor
1 Gaming Floor
1 Courtyard
2 Lockable Dorms

The pod is designed with a desert design in mind, made to look as though it was indeed placed there rather then built there.

It is size 33x33 so when marking out a spot to place it go to where you want any corner, then go in a 17x17. So if standing where you want the South-West corner to be, walk to the East 16 tiles (Not including the one you were standing on), then North 16 tiles (same here). I'm going to try making my pods uneven sizes like this so it'll be easy to spawn them.

Any comments, bug reports, or other things you are welcome to contact me, SlaysInDutch on Discord.