Disabling bellysprite plane for now.
Disabling the separate plane for bellysprites to unfuck the layering for the time being until a better method is figured out. This is not a complete revert and the code still exists, but for the time being the pref setting to hide other people's bellysprites won't be doing much.

If a person who cannot play here with bellysprites being visible really exists and has a problem with this, they are free to submit a github issue for it or figure out a way to implement it without causing issues for the rest of the community.