Makes limbs able to be chopped off with reasonable limits (#16766)
Moves around limb code a bit to make cutting/destroying limbs ACTUALLY possible.

Previously, destroying a limb required you to:
1. Reach the person's max_damage to even let you start having a _chance_ to destroy the limb.
2. Have to have a REALLY strong weapon to destroy/cut limbs. We're talking about a weapon requiring 50 to 80 damage depending on the limb to destroy it.
3. Even IF you managed to get a weapon that could destroy a limb, you STILL had a random prob() based on the damage dealt in that blow.

This changes it so instead of requiring you to reach the limb's max_damage to start having chances to drop, it makes it so limbs have two factors: How damaged the limb currently is AND how strong the blow it.

This still has thresholds. You can't grab a 5 damage butterknife and chop people's limbs off. But if you step in a 30 damage beatrap, or have a 45 damage rock drop on your hand, it might actually get gibbed or knocked off!