Digitigrade for hanner and replicants (#16795)
Added the option for hanner and replicants to use digitigrade legs.

Changed the total reform ability to account for clothing to match the sprites to the leg type of the new form.

Why not proteans and prommies? Proteans use prosthetic legs that do not change appearance based on the digitigrade option. Prometheans do work with digitigrade legs but they currently do not match the opacity of the rest of the body.