Fixes food replicator math error (and buffs it a little bit)
The food replicator had an issue where the amount of nutriment needed to print an item depended on how far it was from half of its maximum volume. I figured that was way too overcomplicated and simply made the cost be based on the volume of reagents in the dish.
I also made the efficiency, which is a multiplier to the amount of nutriment used, be equal to 3 divided by the manipulator's tier. This way, at any level below 3, it will cost more nutriment to make the food than the food has. At 3 it will be equal, and above 3, it will be even cheaper. The reasoning behind this is that nutriment is incredibly easy to get. The largest limiting factor is that you have to grind up whatever has nutriment in order to use it in the first place. Letting the nutriment cost get this low helps the food replicator actually serve as a scene tool.