Appreciation Day (#17427)
* Appreciation Day

A new little system to create and announce Appreciation Days for certain species, and a few other funny things.

The first announcement is 90 minutes into the shift, where a random species will be chosen. This species is one from either a long list, or one of the crew's own species! This is then remembered and further announcements (on average less than every two hours) will use the same thing to appreciate for the rest of that shift.

It seems suitably horribly corporate to me, with some potential for comedy.

* Update appreciation.dm

* Terrible factoids

* Update appreciation.dm

* Update appreciation.dm

* Fixes some things

* unfuck that

* .

* .

* .

* cut it

* fix it

* .

* Set the test values back to proper values

* Update appreciation.dm

* resume

---------

Co-authored-by: Kashargul <144968721+Kashargul@users.noreply.github.com>