Default Save Organs/Markings to No (#17452)
Changed the default setting for Save Organs and Save Markings to No in character setup, which seem to cause more problems for people than people who use them intentionally. This won't change any existing character's settings.

For clarity for those unsure, these are the settings that decide that when you leave the round, whether changes to your organs or markings persist into future rounds. This is can be cool from a story perspective, but on a server like this the majority of people tend to want more control over their own characters and having settings change like this unexpectedly (sometimes due to bugs) is a pain.