Expanded leisure area on the tether (#17441)
* Expanded leisure area on the tether

Significant changes to the tether's pool area. The tiny, highly exposed gym has been moved to the south of the pool, therefore a bit more secluded but also larger and prettied up. The neighbouring toilets now have a locker room. There is a small session room that can be locked on the south side of the gym. There is a very small sauna and massage room on the west of the pool. There are now a few tables and chairs around the pool now, and the old gym has been converted into another, more visible pool-themed cafe/bar (no alcohol!)
Also added treadmills.

* Wires and non-smoking area

* Conflict fix

* Revert "Conflict fix"

This reverts commit e8b98f8c7aef0a310b52ab6a5fb814cbc2a02aff.

* redo conflict fix