Removes sleeping from the code in more places (#17418)
* Axes unused 'paused' variable from projectile code

* Remove TGUI in scrubber init

Moves it to only happen when a user interacts with it.

* Ghostpod tgui input from init -> LateInit

Ghost query REQUIRES a tgui input. It can't be fixed on that level.
Ghost pods however created a ghost query when spawned if they are NOT mapspawn. This means it can only be moved to lateinit, as far as I can see

* Moves borer ghost query from init to lateinit

* Better mob chasing. Gets rid of sleep in mob chasing

Moves Noodle's and Ian's snack tracking var to base /simple_mob
Moves Noodle's and Ian's tracking process from a individual proc of each to /simple_mob
Moves the sleeps out from the tracking proc and makes it a repeating timer.
Noodle now gets nutrition when you feed him.

* Moves blob query to lateinit

Once again, blob HAS to do a ghost query when it's created.
Ghost query will ALWAYS make a tgui window.

* No sleep in cyborg code

Moves pick_module() from /init on robot_module to /LateInit on /robot
photo showing it popped up the module select when I spawned in as a borg: https://i.imgur.com/5rGbqIR.png

* Moves AI controlled human's 'death' proc to LateInitialize

* Moves Corpse's processing from Init to LateInit

death() has a LOT that happens inside of it.

* Let's actually make it lateload on the human itself

* In case we don't want to auto kill ai controlled mobs

* Fixes death() some more

- Protean blobs will ALWAYS have a humanform, so that segment of code was unused.
- Makes the bulk of the exit VR section in a new perform_exit() proc
- Gets rid of the WIP proof of concept hack in there.

* whoops duplicate

* We will ALWAYS have a humanform. ALWAYS.

The server will CRASH if we do not.

* get rid of this...

* Makes corpsespawn work better

* more VR stuff

* No suspending the whole machines subsystem

* avatar

* as anything

* u2

* nolist

* forgot this one

* fix

* VR fix

Fixes an oversight with VR sleepers

* player initiated

* OVERSIGHT

* Update borer.dm

OF course WE DON'T have a mind, we're a GHOST.
Fixes borer spawning

* this was unused

no