[WIP] Custom Links in Character Directory [Advice Requested] (#17790)
* [WIP] Custom Links in Character Directory

Added custom links to the character directory.
They work fine at the moment as just plain text, even if you have to copy and paste them it's an improvement. However, I wanted to keep this in WIP until I have a chance to ask our TGUI experts if it is possible to make them a clickable link.

* maks link usable

* ignore bad links

* .

---------

Co-authored-by: Kashargul <144968721+Kashargul@users.noreply.github.com>