Extra bartender slot (#18505)
Added an extra bartender slot on request.

Reasoning is that it's one of the most social roles and keeps other people moving socially too. However, it's also a very common one for the players to vanish to do scenes, which we don't want to discourage. Instead, letting another person in to fill the role in the meantime is perfect. This already works fine on the tether where there are 3 bars (4 if you include the maint one). SD will still need to share a bit though, or get creative, not really a biggie.