Negative trait point changes. (#18509)
Changed some of the negative traits point values, primarily as they give so many points relative to the small cost of positive traits:
Low Blood Volume from 3 to 2 and the extreme version from 5 to 3. These are serious, but they're reasonably in-line with other serious weakness traits.
Reduced biocompatibility 2->1, 4->2, 8->4. These are really not that much of an issue, if you are in medbay getting treated, it just means they need to stick you with a couple extra syringes and rarely actually causes a problem other than taking a couple more minutes. 4 points is still a lot.
Ambulant Blood 1 to 0, this has no effect at all here where changelings are never a risk.