Shibari! (#18514)
* Shibari!

Added a new scene item, shibari bindings!
These by default only cover the torso, but by clicking on them in your hands, they can also be set to bind arms and legs, acting in a similar way to handcuffs. They are worn in the suit slot.

They can be found in the costume vendors, maint and the restraints crate from cargo.

They have sprites for humanoid mobs and digitigrade legs, but not tesh yet.

* Fix mistake

Fixes error in previous commit

* Update code/game/objects/items/weapons/handcuffs.dm

Co-authored-by: ShadowLarkens <shadowlarkens@gmail.com>

* Update code/modules/mob/living/carbon/human/human.dm

Co-authored-by: ShadowLarkens <shadowlarkens@gmail.com>

* Make these defines to prevent copy paste errors

---------

Co-authored-by: ShadowLarkens <shadowlarkens@gmail.com>
Co-authored-by: Cameron Lennox <killer65311@gmail.com>