Aegis Update
The goal of this PR is to update the Aegis carrier to have a proper AI core and camera network.

Secondary goal, make people actually like going onboard the Aegis.

TODO:
Full camera network.

COMPLETED:
Bluespace Gigabeacon (to allow command / synths access to the carrier in the event of a power outage)

AI core and AI access hallway, with nanite tiles and turrets.